+++ multimno/core/data_objects/silver/silver_present_population_data_object.py
@@ -9,7 +9,7 @@ from pyspark.sql.types import (
     ByteType,
     ShortType,
     TimestampType,
-    LongType,
+    IntegerType,
 )
 
 from multimno.core.data_objects.data_object import ParquetDataObject
@@ -24,7 +24,7 @@ class SilverPresentPopulationDataObject(ParquetDataObject):
     ID = "SilverPresentPopulationDO"
     SCHEMA = StructType(
         [
-            StructField(ColNames.grid_id, LongType(), nullable=False),
+            StructField(ColNames.grid_id, IntegerType(), nullable=False),
             StructField(ColNames.population, FloatType(), nullable=False),
             StructField(ColNames.timestamp, TimestampType(), nullable=False),
             # partition columns
